tools/img2rgb/requirements.txt
pillow>=10.0.0

To install or upgrade Pillow, run the following command:

pip install pillow
